+++ src/day7.rs
@@ -3,7 +3,7 @@ use rayon::prelude::*;
 
 const INPUT: &str = "3,8,1001,8,10,8,105,1,0,0,21,38,47,72,97,122,203,284,365,446,99999,3,9,1001,9,3,9,1002,9,5,9,1001,9,4,9,4,9,99,3,9,102,3,9,9,4,9,99,3,9,1001,9,2,9,102,5,9,9,101,3,9,9,1002,9,5,9,101,4,9,9,4,9,99,3,9,101,5,9,9,1002,9,3,9,101,2,9,9,102,3,9,9,1001,9,2,9,4,9,99,3,9,101,3,9,9,102,2,9,9,1001,9,4,9,1002,9,2,9,101,2,9,9,4,9,99,3,9,1001,9,2,9,4,9,3,9,101,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,1001,9,1,9,4,9,3,9,102,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,1,9,4,9,3,9,101,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,1001,9,2,9,4,9,99,3,9,1001,9,1,9,4,9,3,9,101,1,9,9,4,9,3,9,101,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,101,1,9,9,4,9,3,9,102,2,9,9,4,9,3,9,102,2,9,9,4,9,99,3,9,1001,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,1001,9,2,9,4,9,3,9,102,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,1,9,4,9,3,9,101,1,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,102,2,9,9,4,9,99,3,9,101,1,9,9,4,9,3,9,101,1,9,9,4,9,3,9,102,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,2,9,4,9,3,9,1001,9,2,9,4,9,3,9,102,2,9,9,4,9,3,9,1001,9,1,9,4,9,99,3,9,101,2,9,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,2,9,4,9,3,9,102,2,9,9,4,9,3,9,102,2,9,9,4,9,3,9,101,1,9,9,4,9,3,9,1002,9,2,9,4,9,3,9,1002,9,2,9,4,9,3,9,101,2,9,9,4,9,3,9,1001,9,2,9,4,9,99";
 
-pub fn solve_day_7_pt_1() -> i32 {
+pub fn solve_day_7_pt_1() -> i64 {
     permutations([0, 1, 2, 3, 4])
         .par_iter()
         .map(|permutation| run_amplifier(permutation))
@@ -11,11 +11,11 @@ pub fn solve_day_7_pt_1() -> i32 {
         .unwrap()
 }
 
-fn run_amplifier(phase_settings: &[u8]) -> i32 {
+fn run_amplifier(phase_settings: &[u8]) -> i64 {
     let mut machine = IntcodeMachine::from(INPUT);
     let mut ret = 0;
     for phase in phase_settings.iter() {
-        machine.step(Some(*phase as i32));
+        machine.step(Some(*phase as i64));
 
         loop {
             if let IntcodeReturns::Val(result) = machine.step(Some(ret)) {
@@ -30,7 +30,7 @@ fn run_amplifier(phase_settings: &[u8]) -> i32 {
     ret
 }
 
-pub fn solve_day_7_pt_2() -> i32 {
+pub fn solve_day_7_pt_2() -> i64 {
     permutations([5, 6, 7, 8, 9])
         .par_iter()
         .map(|permutation| run_amplifier_feedback(permutation))
@@ -38,7 +38,7 @@ pub fn solve_day_7_pt_2() -> i32 {
         .unwrap()
 }
 
-fn run_amplifier_feedback(phase_settings: &[u8]) -> i32 {
+fn run_amplifier_feedback(phase_settings: &[u8]) -> i64 {
     let mut machines = [
         IntcodeMachine::from(INPUT),
         IntcodeMachine::from(INPUT),
@@ -47,7 +47,7 @@ fn run_amplifier_feedback(phase_settings: &[u8]) -> i32 {
         IntcodeMachine::from(INPUT),
     ];
     for (idx, phase) in phase_settings.iter().enumerate() {
-        machines[idx].step(Some(*phase as i32));
+        machines[idx].step(Some(*phase as i64));
     }
     let mut ret = 0;
     loop {
